@@ -13,78 +14,79 @@
   "bugs": {
     "url": "https://github.com/rcaferati/react-awesome-button/issues"
   },
-  "jest": {
-    "setupTestFrameworkScriptFile": "./src/setupTest.js",
-    "testURL": "http://localhost"
-  },
   "scripts": {
-    "start": "webpack-dev-server -d --config webpack.demo.config.js",
+    "start": "webpack-dev-server --config webpack.demo.config.js",
     "build": "webpack --mode production",
+    "build:typescript": "webpack --mode production --config webpack.typescript.config.js",
     "build:scss": "webpack --mode production --config webpack.scss.config.js && rm ./dist/styles.js",
     "build:website": "webpack --mode production --config webpack.website.config.js",
     "build:server": "webpack --mode production --config webpack.server.config.js",
     "build:themes": "babel-node ./theme-builder.js",
-    "test": "BABEL_JEST_STAGE=0 jest",
     "lint": "eslint src/**",
     "prepublish": "npm run build && npm run build:scss && npm run build:themes",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
-  },
-  "peer-dependencies": {
-    "react": "^15.0.1 || ^16.0.0",
-    "prop-types": "^15.0.0"
+    "test": "jest",
+    "storybook": "start-storybook -p 40123 -s public",
+    "build-storybook": "build-storybook -s public --quiet"
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^3.4.10",
-    "@storybook/addon-centered": "^3.4.10",
-    "@storybook/addon-info": "^3.4.10",
-    "@storybook/addon-knobs": "^3.4.10",
-    "@storybook/addon-links": "^3.4.10",
-    "@storybook/addon-storysource": "^3.4.10",
-    "@storybook/addon-viewport": "^3.4.10",
-    "@storybook/addons": "^3.4.10",
-    "@storybook/react": "^3.4.10",
-    "autoprefixer": "^7.0.1",
-    "babel": "^6.23.0",
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.24.1",
-    "babel-eslint": "^8.2.1",
-    "babel-jest": "^22.0.4",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.12",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-runtime": "^6.26.0",
-    "css-loader": "^0.28.1",
-    "enzyme": "^3.9.0",
+    "@babel/core": "^7.19.3",
+    "@babel/node": "^7.19.1",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/preset-env": "^7.19.4",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@storybook/addon-actions": "^6.5.12",
+    "@storybook/addon-essentials": "^6.5.12",
+    "@storybook/addon-interactions": "^6.5.12",
+    "@storybook/addon-knobs": "^6.4.0",
+    "@storybook/addon-links": "^6.5.12",
+    "@storybook/addon-storysource": "^6.5.12",
+    "@storybook/builder-webpack5": "^6.5.12",
+    "@storybook/manager-webpack5": "^6.5.12",
+    "@storybook/preset-scss": "^1.0.3",
+    "@storybook/react": "^6.5.12",
+    "@storybook/testing-library": "^0.0.13",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@types/enzyme": "^3.10.12",
+    "@types/enzyme-adapter-react-16": "^1.0.6",
+    "@types/jest": "^29.2.0",
+    "@types/react": "^18.0.21",
+    "@types/react-test-renderer": "^18.0.0",
+    "autoprefixer": "^10.4.12",
+    "babel-loader": "^8.2.5",
+    "css-loader": "^6.7.1",
+    "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.13.1",
-    "eslint": "^4.15.0",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-config-prettier": "^4.1.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.5.1",
-    "jest": "^22.0.4",
-    "loki": "^0.14.2",
-    "mini-css-extract-plugin": "^0.8.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.3",
-    "postcss-loader": "^2.0.10",
-    "prismjs": "^1.9.0",
-    "prop-types": "^15.6.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
-    "react-test-renderer": "^16.8.6",
-    "sass-loader": "^6.0.3",
-    "shelljs": "^0.7.8",
-    "storybook-addon-smart-knobs": "^3.3.1",
-    "style-loader": "^0.19.1",
-    "terser-webpack-plugin": "^2.1.2",
-    "webpack": "^4.40.2",
-    "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^3.8.1"
+    "eslint": "^8.25.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-react": "^7.31.10",
+    "jest": "^29.2.1",
+    "mini-css-extract-plugin": "^2.6.1",
+    "node-sass": "^7.0.3",
+    "optimize-css-assets-webpack-plugin": "^6.0.1",
+    "postcss-loader": "^7.0.1",
+    "prismjs": "^1.29.0",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "^6.4.2",
+    "react-router-dom": "^6.4.2",
+    "react-test-renderer": "^18.2.0",
+    "sass-loader": "^13.1.0",
+    "shelljs": "^0.8.5",
+    "storybook-addon-sass-postcss": "^0.1.3",
+    "style-loader": "^3.3.1",
+    "terser-webpack-plugin": "^5.3.6",
+    "ts-jest": "^29.0.3",
+    "ts-loader": "^9.4.1",
+    "typescript": "^4.8.4",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.11.1"
   },
   "keywords": [
     "react-component",
