@@ -8,10 +8,7 @@ const Footer = ({ repository, article }) => (
       <img className="support" src="/images/support.svg" alt="Modern Web Browsers" title="Modern Web Browsers" />
     </div>
     <small>
-      Promote and support this project on <a href={repository}>github</a>.
-    </small>
-    <small>
-      Read more and discuss at the <a href={article}>article page</a>.
+      Support this project on <a href={repository}>github</a>
     </small>
     <small>Built by <a href="https://caferati.me"><b>@rcaferati</b></a></small>
   </footer>
