@@ -1,8 +1,8 @@
-import React from 'react';
-import { render } from 'react-dom';
+import * as React from "react";
+import * as ReactDOM from "react-dom";
 import Demo from './demo';
 
-render(
+ReactDOM.render(
   (
     <Demo />
   ),
